refactor: split startup guard tests
